add create_quiz_party and get_quiz_list rpc and repo implementation, some fixes

+++ internal/models/models.go
@@ -1,24 +1,24 @@
 package models
 
 type UserAccount struct {
-	ID   int
+	ID   int64
 	Name string
 }
 
 type Quiz struct {
-	ID   int
+	ID   int64
 	Name string
 }
 
 type Participation struct {
-	ID            int
-	UserAccountID int
-	QuizID        int
+	ID            int64
+	UserAccountID int64
+	QuizID        int64
 }
 
 type ResponseReport struct {
-	ID              int
-	ParticipationID int
+	ID              int64
+	ParticipationID int64
 	Correct         bool
 	PenaltyTime     int
 }
